remove urls
