+++ debug_train_kialo.py
@@ -18,7 +18,7 @@ sys.argv += '--eval_not_trained True --do_train False'.split(' ')
 sys.argv += '--annotated_samples_in_test True --debug_map_index 17763'.split(' ')
 
 # eval annotated samples
-sys.argv += '--do_eval False --do_eval_annotated_samples True --debug_maps_size 3'.split(' ')
+# sys.argv += '--do_eval False --do_eval_annotated_samples True --debug_maps_size 3'.split(' ')
 
 # debug specific map
 # sys.argv += '--no_data_split True --debug_map_index 22187'.split(' ')
